+++ src/Memio/Model/File.php
@@ -20,7 +20,6 @@ use Memio\Model\Phpdoc\LicensePhpdoc;
  */
 class File
 {
-    public string $filename;
     public ?LicensePhpdoc $licensePhpdoc = null;
     public array $fullyQualifiedNames = [];
     public ?Structure $structure = null;
@@ -28,9 +27,8 @@ class File
     /**
      * @api
      */
-    public function __construct(string $filename)
+    public function __construct(public string $filename)
     {
-        $this->filename = $filename;
     }
 
     /**